+++ package-scripts.js
@@ -1,14 +1,28 @@
-const {concurrent, rimraf} = require('nps-utils')
+const {concurrent, rimraf, getBin, series} = require('nps-utils')
 const {version} = require('./package.json')
+
+let jestBin
+try {
+	jestBin = getBin('jest-cli', 'jest')
+} catch (err) {
+	jestBin = 'pleaseInstallJest'
+}
 
 const runBabel = `NODE_ENV=production babel -s true --ignore '**/*.test.js,**/__snapshots__' -d dist/`
 const scripts = {
 	build: {
-		default: `nps build.clean build.babel`,
+		default: `nps build.clean build.babel build.doc`,
 		clean: rimraf('dist/'),
 		babel: `${runBabel} src/`,
 		watch: `${runBabel} --watch src/`,
 		git: `sh build-git.sh v${version.split('.')[0]}`,
+		doc: series(
+			`echo '# API' > API.md`,
+			`echo >> API.md`,
+			`jsdoc2md -f src/*.js src/**/*js >> API.md`,
+			`echo >> API.md`,
+			`git log -n1 --format=format:"_Generated from %H, %cI_" >> API.md`
+		),
 	},
 	test: {
 		default: concurrent.nps('test.lint', 'test.full'),
@@ -20,10 +34,7 @@ const scripts = {
 		},
 		full: 'NODE_ENV=test jest --coverage --color',
 		watch: 'NODE_ENV=test jest --color --watch',
-		inspect: `NODE_ENV=test node --inspect ${
-			// ok, a little bit of a hack
-			require.resolve('jest-cli').replace('build', 'bin')
-		} --runInBand --watch`,
+		inspect: `NODE_ENV=test pnpx ndb ${jestBin} --runInBand --watch`,
 	},
 	publish: `npm publish --access public`,
 }