+++ src/DB/index.js
@@ -1,0 +1,2 @@
+export {default} from './DB'
+export {default as SQLite, valToSql, sql} from './SQLite'