+++ src/EventSourcingDB/index.js
@@ -1,1 +1,3 @@
 export {default} from './EventSourcingDB'
+export {default as applyResult} from './applyResult'
+export {default as ESModel} from './ESModel'